Fix issue clawpack/visclaw#118
`add_colormap` now constructs a colormap that is strictly diverging
between the two given colormaps, i.e. the first colormap goes from 0-0.5
and the second from 0.5-1.0.  It also now returns a Normalize object
that needs to be handed to the appropriate plotting command to map the
data_limits and data_break correctly to the colormap.