Changes to support plotting fgout grids as frames, more flexible in general
This makes it possible to specify a different file_prefix rather than fort
for the output files fort.t*, fort.q*, etc.
Calling sequences of a few functions needed new arguments to support this.
Some files in resulting _plots directory renamed in general to be simpler,
in particular the jsanimation and 'all frames' files.