Add meson build. (#293)
Only installs python files for now.

This is the first phase of switching to meson (away from distutils). This allows meson to install python files only. The second phase is to add compilation of the Fortran extensions (for Python).

For now, this will have no effect since the meson build is not enabled at the top level.