update geoplot.py to handle 1d or 2d q array, with eta in q[-1,...]
Use ellipses q[0,...] for depth, which is q[0,:] in 1d or q[0,:,:] in 2d.
Use eta = q[-1,...] to work in 1d or 2d and also with 2d bouss where
q will have additional components for hu and hv correction terms.
Assumes that eta is always the final component of the q array read in
from geoclaw output.